--- Assets/Scripts/Player/PlayerInput.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: d3fe3d00ac4c25d44a7bb90463d7d561
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
